Se agrego la autorización de modificar si es propia del usuario

+++ server/src/controller/team.controller.js
@@ -7,8 +7,8 @@ exports.saveEquipos = async(req, res) =>{
     try{
     const params = req.body; 
     const data = {
-        name:params.name
-        
+        name:params.name,
+        user: req.user.sub
     }   
     const  msg = validate.validateData(data);
     if(msg) return res.status(400).send(msg); 
@@ -27,7 +27,10 @@ exports.saveEquipos = async(req, res) =>{
 
 exports.equiposDelete = async(req,res)=>{
     try{
-        const equipoId = req.params.id; 
+        const equipoId = req.params.id;
+        const team = await Equipos.findOne({_id: equipoId}).lean();
+        const account = await validate.ownAccount(team.user, req.user.sub);
+        if(account ) return res.send(account) 
         const delEquipos = await Equipos.findOneAndDelete({_id: equipoId}); 
         if(!delEquipos) return res.status(404).send({message: 'Equipos not found or already deleted'}); 
         return res.send({message: 'Equipos Deleted', delEquipos});
@@ -42,13 +45,16 @@ exports.equiposUpdate = async(req,res)=>{
     try{
         const params = req.body; 
         const equiposId = req.params.id; 
+        const team = await Equipos.findOne({_id: equiposId}).lean();
+        const account = await validate.ownAccount(team.user, req.user.sub);
+        if(account ) return res.send(account) 
         const check = validate.checkUpdate(params);
         if(check === false) return res.status(400).send('data is not recivida'); 
         const equiposUpdate = await Equipos.findOneAndUpdate({_id: equiposId}, params, {new: true});
         return res.send({message: 'Category update', equiposUpdate});
     }catch(err){
         console.log(err);
-        return res.status(500).send({err, message: 'Error deleting'});
+        return res.status(500).send({err, message: 'Error updating'});
     }
 }
